почти весь функционал сделан. Осталась только лента новостей

+++ news/views.py
@@ -34,8 +34,8 @@ def profile(request):
 @login_required
 def blogs(request):
     user = request.user
-    blog_list = models.NewsItem.objects.filter(author=user).order_by('-pub_date')[:3]
-    comment_list = models.Comment.objects.filter(author=user).order_by('-pub_date')[:3]
+    blog_list = models.NewsItem.objects.filter(author=user).order_by('-pub_date')
+    comment_list = models.Comment.objects.filter(author=user).order_by('-pub_date')
     return render(request, 'news/blogs/blogs.html',
                   {'blogs': blog_list, 'comments': comment_list})
 
@@ -60,10 +60,12 @@ def create_blog(request):
 
 @login_required
 def delete_blog(request, pk):
-    # Получаем новость по ID
+    user = request.user
     blog = get_object_or_404(models.NewsItem, pk=pk, author=request.user)  # Только для авторизованного пользователя
-    blog.delete()  # Удаляем новость
-    return redirect('news:blogs')
+    if user.id == blog.author.id:
+        blog.delete()
+        return redirect('news:blogs')
+    render(request, '400.html', {'message': 'Вы не можете удалять чужую новость'})
 
 
 @login_required
@@ -74,9 +76,47 @@ def edit_blog_form(request, pk):
 
 @login_required
 def edit_blog(request, pk):
-    return redirect('news:blogs')
+    if request.method == 'POST':
+        user = request.user
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        image = request.FILES.get('image')
+        blog = models.NewsItem.objects.get(pk=pk)
+        if blog.author == user:
+            blog.title = title
+            blog.content = content
+            if image:
+                blog.picture = image
+            blog.save()
+            return redirect('news:blogs')
+        render(request, '400.html', {'message': 'Вы не можете редактировать эту новость'})
+    render(request, '400.html', {'message': 'Только POST запросы разрешены для редактирования новости'})
 
 
 def view_blog(request, pk):
     blog = get_object_or_404(models.NewsItem, pk=pk)
-    return render(request, 'news/blogs/view_blog.html', {'blog': blog})
+    comments = blog.comments.all().order_by('-pub_date')
+    return render(request, 'news/blogs/view_blog.html', {'blog': blog, 'comments': comments}, )
+
+
+@login_required
+def create_comment(request, pk):
+    if request.method == 'POST':
+        content = request.POST.get('content')
+        image = request.FILES.get('image')
+        user = request.user
+        blog = models.NewsItem.objects.get(pk=pk)
+        comment = models.Comment.objects.create(author=user, content=content, picture=image, news_item=blog)
+        comment.save()
+        return redirect('news:view_blog', pk=pk)
+    render(request, '400.html', {'message': 'Только POST запросы разрешены для создания комментариев'})
+
+
+@login_required
+def delete_comment(request, pk):
+    user = request.user
+    comment = get_object_or_404(models.Comment, pk=pk)
+    if user == comment.author:
+        comment.delete()
+        return redirect('news:blogs')
+    render(request, '400.html', {'message': 'Вы не можете удалять чужой комментарий'})
